ai-services/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
python-dotenv==1.0.0
google-generativeai==0.3.2
sentence-transformers==2.2.2
scikit-learn==1.3.2
pandas==2.1.4
numpy==1.25.2
scipy==1.11.4
wikipediaapi==0.6.0
requests==2.31.0
python-multipart==0.0.6
aiofiles==23.2.1
httpx==0.25.2
pytest==7.4.3
pytest-asyncio==0.21.1
